// src/interfaces/Planet.ts
export interface Planet {
	id: string;
	name: string;
	rotation_period: string;
	orbital_period: string;
	diameter: string;
	climate: string;
	gravity: string;
	terrain: string;
	surface_water: string;
	population: string;
	city_name: string;
	city_region: string;
	city_country: string;
	city_lat: number;
	city_lon: number;
	city_tz_id: string;
	city_temp_c: number;
	city_temp_f: number;
	city_humidity: number;
};